+++ app/src/main/java/com/hedvig/app/feature/insurance/ui/InsuranceAdapter.kt
@@ -303,128 +303,3 @@ class InsuranceAdapter(
         private const val SUBHEADING = 2
     }
 }
-
-@Composable
-fun Subheading(text: String) {
-    Text(
-        text = text,
-        style = MaterialTheme.typography.h6,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(
-                start = 16.dp,
-                top = 48.dp,
-                end = 16.dp,
-                bottom = 8.dp,
-            ),
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun SubheadingPreview() {
-    HedvigTheme {
-        Subheading("Add more coverage")
-    }
-}
-
-/*
- * Note: This Composable uses hardcoded colors due to difficulties with
- * declaring a particular component to be in dark theme instead of the
- * default. When we update `HedvigTheme` to be Compose-first instead of
- * XML-Theme first, we can reconfigure the theme for this composable to
- * be `dark` no matter what the system value is.
- */
-@Composable
-fun CrossSell(
-    data: InsuranceModel.CrossSell,
-    onCtaClick: () -> Unit,
-) {
-    val placeholder by rememberBlurHash(data.backgroundBlurHash, 64, 32)
-    Card(
-        border = BorderStroke(1.dp, hedvigBlack12percent),
-        modifier = Modifier
-            .padding(
-                horizontal = 16.dp,
-                vertical = 8.dp,
-            )
-            .height(200.dp),
-    ) {
-        Image(
-            painter = rememberImagePainter(
-                data = data.backgroundUrl,
-                builder = {
-                    transformations(CropTransformation())
-                    placeholder(placeholder)
-                    crossfade(true)
-                },
-            ),
-            contentDescription = null,
-            contentScale = ContentScale.Crop,
-            modifier = Modifier.fillMaxSize(),
-        )
-        Box(
-            contentAlignment = Alignment.BottomCenter,
-            modifier = Modifier
-                .background(
-                    brush = Brush.verticalGradient(
-                        colors = listOf(
-                            Color(0x00000000),
-                            Color(0xFF000000),
-                        ),
-
-                    )
-                )
-                .padding(16.dp),
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.Bottom,
-                modifier = Modifier.fillMaxWidth(),
-            ) {
-                Column(modifier = Modifier.padding(bottom = 4.dp)) {
-                    Text(
-                        text = data.title,
-                        style = MaterialTheme.typography.subtitle1,
-                        color = whiteHighEmphasis,
-                    )
-                    Text(
-                        text = data.description,
-                        style = MaterialTheme.typography.subtitle2,
-                        color = whiteHighEmphasis,
-                        modifier = Modifier.padding(top = 4.dp),
-                    )
-                }
-                Button(
-                    onClick = onCtaClick,
-                    colors = ButtonDefaults.buttonColors(
-                        backgroundColor = whiteHighEmphasis,
-                        contentColor = hedvigBlack,
-                    )
-                ) {
-                    Text(
-                        text = data.callToAction,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun CrossSellPreview() {
-    HedvigTheme {
-        CrossSell(
-            data = InsuranceModel.CrossSell(
-                title = "Accident Insurance",
-                description = "179 kr/mo.",
-                callToAction = "Calculate price",
-                action = InsuranceModel.CrossSell.Action.Chat,
-                backgroundUrl = "https://images.unsplash.com/photo-1628996796855-0b056a464e06",
-                backgroundBlurHash = "LJC6\$2-:DiWB~WxuRkayMwNGo~of",
-            ),
-            onCtaClick = {}
-        )
-    }
-}
